+++ src/lib/abt_did_workshop/application.ex
@@ -4,22 +4,10 @@ defmodule AbtDidWorkshop.Application do
   alias AbtDidWorkshopWeb.Endpoint
 
   def start(_type, _args) do
-    filename = :abt_did_workshop |> Application.app_dir() |> Path.join("priv/forge.toml")
-    servers = ForgeSdk.init(:cert, "", filename)
-
-    children =
-      servers ++
-        [
-          AbtDidWorkshopWeb.Endpoint,
-          AbtDidWorkshop.UserDb,
-          AbtDidWorkshop.AssetsDb,
-          AbtDidWorkshop.AppState
-        ]
-
+    children = get_children()
     register_type_urls()
     opts = [strategy: :one_for_one, name: AbtDidWorkshop.Supervisor]
     result = Supervisor.start_link(children, opts)
-    init_certs()
     result
   end
 
@@ -36,9 +24,22 @@ defmodule AbtDidWorkshop.Application do
     ])
   end
 
-  defp init_certs do
-    {robert, _} = AbtDidWorkshop.WalletUtil.init_robert()
-    Process.sleep(5000)
-    AbtDidWorkshop.AssetUtil.init_certs(robert, "ABT", 40)
+  def get_children() do
+    app_servers1 = [
+      AbtDidWorkshopWeb.Endpoint,
+      AbtDidWorkshop.UserDb,
+      AbtDidWorkshop.AssetsDb,
+      AbtDidWorkshop.Repo
+    ]
+
+    app_servers2 = [AbtDidWorkshop.AppState]
+
+    filename = :abt_did_workshop |> Application.app_dir() |> Path.join("priv/forge.toml")
+    forge_servers = ForgeSdk.init(:abt_did_workshop, "", filename)
+
+    case Application.get_env(:abt_did_workshop, :env) do
+      "test" -> app_servers1
+      _ -> forge_servers ++ app_servers1 ++ app_servers2
+    end
   end
 end