+++ src/lib/abt_did_workshop/models/tx_behavior.ex
@@ -1,0 +1,26 @@
+defmodule AbtDidWorkshop.TxBehavior do
+  @moduledoc """
+  Represents the datastructure for demo.
+  """
+
+  use Ecto.Schema
+
+  import Ecto.Changeset
+
+  alias AbtDidWorkshop.Tx
+
+  schema("tx_behavior") do
+    field(:tx_type, :string)
+    field(:behavior, :string)
+    field(:token, :integer)
+    field(:asset, :string)
+    field(:function, :string)
+    belongs_to(:tx, Tx)
+  end
+
+  def changeset(struct, params \\ %{}) do
+    struct
+    |> cast(params, [:tx_type, :behavior, :token, :asset, :function])
+    |> validate_required([:tx_type, :behavior])
+  end
+end