+++ src/lib/abt_did_workshop/utils/util.ex
@@ -54,9 +54,21 @@ defmodule AbtDidWorkshop.Util do
     "#{path}?appPk=#{app_pk}&appDid=#{app_state.did}&action=requestAuth&url=#{url}"
   end
 
+  def gen_deeplink(tx) do
+    url = (get_callback() <> "transaction/#{tx.id}") |> URI.encode_www_form()
+    app_state = AppState.get()
+    path = String.trim_trailing(app_state.path, "/")
+    app_pk = Multibase.encode!(app_state.pk, :base58_btc)
+    "#{path}?appPk=#{app_pk}&appDid=#{app_state.did}&action=requestAuth&url=#{url}"
+  end
+
   def hex_to_bin("0x" <> hex), do: hex_to_bin(hex)
   def hex_to_bin(hex), do: Base.decode16!(hex, case: :mixed)
 
   def did_to_address("did:abt:" <> address), do: address
   def did_to_address(address), do: address
+
+  def empty?(nil), do: true
+  def empty?(""), do: true
+  def empty?(_), do: false
 end