+++ src/lib/abt_did_workshop_web/controllers/page_controller.ex
@@ -1,7 +1,0 @@
-defmodule AbtDidWorkshopWeb.PageController do
-  use AbtDidWorkshopWeb, :controller
-
-  def index(conn, _params) do
-    render(conn, "index.html")
-  end
-end