+++ src/lib/abt_did_workshop_web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule AbtDidWorkshopWeb.PageView do
-  use AbtDidWorkshopWeb, :view
-end