+++ src/priv/repo/migrations/20190303192741_add_function.exs
@@ -1,0 +1,9 @@
+defmodule AbtDidWorkshop.Repo.Migrations.AddFunction do
+  use Ecto.Migration
+
+  def change do
+    alter table(:tx_behavior) do
+      add(:function, :string)
+    end
+  end
+end