+++ src/test/abt_did_workshop_web/views/layout_view_test.exs
@@ -1,0 +1,3 @@
+# defmodule AbtDidWorkshopWeb.LayoutViewTest do
+#   use AbtDidWorkshopWeb.ConnCase, async: true
+# end